Added problem 17.
It's split up into a few functions, one a generic
function that generates a string for a number:
e.g, 42326343 => "forty-two million three hundred
and twenty-six thousand three hundred and three"

To solve, call solve(min, max). It moves from
min to max, uses the letterify function,
and strips spaces and dashes and adds the length.
It returns the length of the combined string
(minus spaces/dashes) of the numbers from min to max